+++ src/main/java/de/uni_hannover/sra/minimax_simulator/io/exporter/csv/SignalHtmlExporter.java
@@ -4,17 +4,14 @@ import de.uni_hannover.sra.minimax_simulator.io.IOUtils;
 import de.uni_hannover.sra.minimax_simulator.model.signal.*;
 import de.uni_hannover.sra.minimax_simulator.model.signal.jump.Jump;
 
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.Writer;
+import java.io.*;
 
 /**
  * The {@code SignalHtmlExporter} exports the {@link SignalTable} to an HTML file.
  *
  * @author Martin L&uuml;ck
+ * @author Philipp Rohde
  */
-// TODO: upgrade to HTML5 standards
 public class SignalHtmlExporter extends AbstractSignalExporter {
 
 	/**
@@ -33,64 +30,64 @@ public class SignalHtmlExporter extends AbstractSignalExporter {
 		Writer wr = null;
 
 		try {
-			wr = IOUtils.toBufferedWriter(new FileWriter(file));
+			// read template file
+			InputStream is = this.getClass().getResourceAsStream("/html/signaltable.html");
+			String template = IOUtils.getStringFromInputStream(is);
 
-			wr.append("<!doctype html>\n" +
-					"<html>" +
-					"<head><meta charset=\"utf-8\"></head><body><table>\n");
-
+			StringBuilder sb = new StringBuilder();
 			// header line
-			wr.append("<tr><th>#</th><th>Label</th>");
+			sb.append("<table><tr><th>#</th><th>Label</th>");
 			for (SignalType signal : config.getSignalTypes()) {
-				wr.append("<th>").append(signal.getId()).append("</th>");
+				sb.append("<th>").append(signal.getId()).append("</th>");
 			}
-			wr.append("<th>Alu == 0?</th><th>Jump</th><th>Description</th></tr>");
+			sb.append("<th>Alu == 0?</th><th>Jump</th><th>Description</th></tr>");
 
 			// value lines
 			for (int i = 0, n = table.getRowCount(); i < n; i++) {
 				SignalRow row = table.getRow(i);
-				wr.append("<tr><td>") .append(Integer.toString(i)).append("</td>");
+				sb.append("<tr><td>").append(Integer.toString(i)).append("</td>");
 				if (row.getLabel() != null) {
-					wr.append("<td>").append(row.getLabel()).append("</td>");
+					sb.append("<td>").append(row.getLabel()).append("</td>");
 				}
 				else {
-					wr.append("<td></td>");
+					sb.append("<td></td>");
 				}
 
 				for (SignalType signal : config.getSignalTypes()) {
 					SignalValue value = row.getSignal(signal);
-					wr.append("<td>").append(getShortDescription(value, signal)).append("</td>");
+					sb.append("<td>").append(getShortDescription(value, signal)).append("</td>");
 				}
 
 				String targets;
 				String cond;
+				String openingTag = "<td>";
 				Jump j = row.getJump();
-
-				if (j.getTargetRow(i, 0) == j.getTargetRow(i, 1)) {
-					cond = "";
-				}
-				else {
-					cond = "1<br/>0";
-				}
-
 
 				int target0 = j.getTargetRow(i, 0);
 				int target1 = j.getTargetRow(i, 1);
+
 				if (target0 == target1) {
+					cond = "";
 					targets = Integer.toString(target0);
 					if (target0 == i + 1) {
-						targets = "<span style=\"color:#777777;\">" + targets + "</span>";
+						openingTag = "<td class=\"defaultJump\">";
 					}
 				}
 				else {
-					targets = target1 + "<br/>" + target0;
+					cond = "1<br>0";
+					targets = target1 + "<br>" + target0;
 				}
 
-				wr.append("<td>").append(cond).append("</td><td>").append(targets).append("</td><td>");
-				wr.append(row.getDescription().replaceAll("\n", "<br/>"));
-				wr.append("</td></tr>");
+				sb.append("<td>").append(cond).append("</td>").append(openingTag).append(targets).append("</td><td>");
+				sb.append(row.getDescription().replaceAll("\n", "<br>"));
+				sb.append("</td></tr>");
 			}
-			wr.append("</table></body></html>");
+			sb.append("</table></body></html>");
+			String signalTable = sb.toString();
+
+			wr = IOUtils.toBufferedWriter(new FileWriter(file));
+			// replace placeholder with content
+			wr.append(template.replace("$table", signalTable));
 		} finally {
 			IOUtils.closeQuietly(wr);
 		}